+++ Source/Engine/UI/UIWidget.h
@@ -35,7 +35,7 @@ public:
 	Rect(float _x, float _y, float _w, float _h) : x(_x), y(_y), w(_w), h(_h) {}
 
 	Rect &Intersect(const Rect &r);
-	Rect &Translate(int _x, int _y) {
+	Rect &Translate(float _x, float _y) {
 		x += _x;
 		y += _y;
 		return *this;
